requirements.txt
beautifulsoup4==4.12.3
GitPython==3.1.43
langchain-core==0.2.0
langchain-openai==0.1.7
python-dotenv==1.0.1
aiohttp~=3.9.5
requests~=2.31.0